# AccountMs/src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Account Microservice API
  description: API REST para gestionar cuentas bancarias del sistema bancario XYZ
  version: 1.0.0
  contact:
    name: NTTDATA Tech Girls Power
    email: support@bancoXYZ.com

servers:
  - url: http://localhost:8082
    description: Servidor de desarrollo

paths:
  /api/v1/accounts:
    post:
      summary: Crear una nueva cuenta bancaria
      description: Crea una cuenta bancaria para un cliente existente
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '201':
          description: Cuenta creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccount'
        '400':
          description: Datos de entrada inválidos
        '500':
          description: Error interno del servidor

    get:
      summary: Listar todas las cuentas bancarias
      description: Obtiene la lista de todas las cuentas bancarias
      operationId: getAllAccounts
      responses:
        '200':
          description: Lista de cuentas obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BankAccount'

  /api/v1/accounts/{id}:
    get:
      summary: Obtener cuenta por ID
      description: Obtiene los detalles de una cuenta específica
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Cuenta encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccount'
        '404':
          description: Cuenta no encontrada

    delete:
      summary: Eliminar cuenta bancaria
      description: Elimina una cuenta bancaria del sistema
      operationId: deleteAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Cuenta eliminada exitosamente
        '404':
          description: Cuenta no encontrada

  /api/v1/accounts/{id}/deposit:
    put:
      summary: Realizar depósito
      description: Deposita dinero en una cuenta bancaria
      operationId: deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmountRequest'
      responses:
        '200':
          description: Depósito realizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccount'
        '400':
          description: Monto inválido
        '404':
          description: Cuenta no encontrada

  /api/v1/accounts/{id}/withdrawal:
    put:
      summary: Realizar retiro
      description: Retira dinero de una cuenta bancaria
      operationId: withdraw
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmountRequest'
      responses:
        '200':
          description: Retiro realizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccount'
        '400':
          description: Saldo insuficiente o monto inválido
        '404':
          description: Cuenta no encontrada

  /api/v1/accounts/clients/{clientId}:
    get:
      summary: Obtener cuentas por cliente
      description: Obtiene todas las cuentas de un cliente específico
      operationId: getAccountsByClientId
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Cuentas del cliente obtenidas exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BankAccount'

components:
  schemas:
    BankAccount:
      type: object
      required:
        - accountId
        - accountNumber
        - balance
        - accountType
        - clientId
      properties:
        accountId:
          type: integer
          format: int64
          description: Identificador único de la cuenta
          example: 1
        accountNumber:
          type: string
          description: Número de cuenta generado automáticamente
          example: "ACC-A1B2C3D4"
        balance:
          type: number
          format: double
          description: Saldo actual de la cuenta
          example: 1500.50
        accountType:
          $ref: '#/components/schemas/AccountType'
        clientId:
          type: integer
          format: int64
          description: ID del cliente propietario
          example: 100

    AccountRequest:
      type: object
      required:
        - clientId
        - accountType
      properties:
        clientId:
          type: integer
          format: int64
          description: ID del cliente
          example: 100
        accountType:
          $ref: '#/components/schemas/AccountType'

    AmountRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          format: double
          minimum: 0.01
          description: Monto para depósito o retiro
          example: 250.75

    AccountType:
      type: string
      enum:
        - SAVINGS
        - CHECKING
      description: |
        Tipo de cuenta bancaria:
        * SAVINGS - Cuenta de ahorros (saldo no puede ser negativo)
        * CHECKING - Cuenta corriente (permite sobregiro hasta -500)